small refactor to provider file structure
