+++ src/providers/Fractal/types/index.ts
@@ -1,7 +1,7 @@
 export * from './daoLegacy';
 export * from './state';
 export * from './treasury';
-export * from './governance';
+export * from '../governance/types';
 export * from './usul';
 export * from './actions';
 export * from './gnosis';
